project/requirements-chat.txt

langchain>=0.1.0,<0.2.0
langchain-core>=0.1.0,<0.2.0
langchain-community>=0.0.10,<0.1.0
langchain-openai>=0.0.5
requests>=2.31.0
pydantic>=2.5.0
python-dotenv>=1.0.0